change EL84 grid leak resistor to 1M
In the day time, I sort of feel the amp is too quiet. That adding the 5.6k resistor inline has reduced the signal. This small resistance divide is like 98% the input. But that is still the 2% loss of signal. I think the gain stage of these preamp tubes is not very strong. And I read the EL84 does have a wider response to small changes in input impedance.